Prefer installed scientific dependencies for paper lab
